refactor: 프로젝트 구조 변경(exception) - 에러코드는 constant 패키지를 만들어 저장함
